SQL Server support

+++ src/Storefront.Menu.API/Migrations/20190909115049_CreateFunctionalIndexes.cs
@@ -9,33 +9,39 @@ namespace Storefront.Menu.API.Migrations
     {
         protected override void Up(MigrationBuilder migrationBuilder)
         {
-            migrationBuilder.Sql($@"
-                CREATE INDEX idx_item_group_name
-                ON {ApiDbContext.Schema}.item_group (public.ci_ai(title));
-            ");
+            if (migrationBuilder.ActiveProvider == "Npgsql.EntityFrameworkCore.PostgreSQL")
+            {
+                migrationBuilder.Sql($@"
+                    CREATE INDEX idx_item_group_name
+                    ON {ApiDbContext.Schema}.item_group (public.ci_ai(title));
+                ");
 
-            migrationBuilder.Sql($@"
-                CREATE INDEX idx_item_name
-                ON {ApiDbContext.Schema}.item (public.ci_ai(name));
-            ");
+                migrationBuilder.Sql($@"
+                    CREATE INDEX idx_item_name
+                    ON {ApiDbContext.Schema}.item (public.ci_ai(name));
+                ");
 
-            migrationBuilder.Sql($@"
-                CREATE INDEX idx_option_group_name
-                ON {ApiDbContext.Schema}.option_group (public.ci_ai(title));
-            ");
+                migrationBuilder.Sql($@"
+                    CREATE INDEX idx_option_group_name
+                    ON {ApiDbContext.Schema}.option_group (public.ci_ai(title));
+                ");
 
-            migrationBuilder.Sql($@"
-                CREATE INDEX idx_option_name
-                ON {ApiDbContext.Schema}.option (public.ci_ai(name));
-            ");
+                migrationBuilder.Sql($@"
+                    CREATE INDEX idx_option_name
+                    ON {ApiDbContext.Schema}.option (public.ci_ai(name));
+                ");
+            }
         }
 
         protected override void Down(MigrationBuilder migrationBuilder)
         {
-            migrationBuilder.Sql(@"DROP INDEX IF EXISTS idx_item_group_name;");
-            migrationBuilder.Sql(@"DROP INDEX IF EXISTS idx_item_name;");
-            migrationBuilder.Sql(@"DROP INDEX IF EXISTS idx_option_group_name;");
-            migrationBuilder.Sql(@"DROP INDEX IF EXISTS idx_option_name;");
+            if (migrationBuilder.ActiveProvider == "Npgsql.EntityFrameworkCore.PostgreSQL")
+            {
+                migrationBuilder.Sql(@"DROP INDEX IF EXISTS idx_item_group_name;");
+                migrationBuilder.Sql(@"DROP INDEX IF EXISTS idx_item_name;");
+                migrationBuilder.Sql(@"DROP INDEX IF EXISTS idx_option_group_name;");
+                migrationBuilder.Sql(@"DROP INDEX IF EXISTS idx_option_name;");
+            }
         }
     }
 }
